Use search params inside useUsers hook
