+++ src/entities/users/ui/users-table.ui.tsx
@@ -1,112 +1,33 @@
 import { ChevronRight as ChevronRightIcon, MoreVert as DotsIcon } from '@mui/icons-material';
 import { IconButton, Menu, MenuItem } from '@mui/material';
 import { Link } from '@tanstack/react-router';
-import fuzzysort from 'fuzzysort';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useAuth } from '~/app';
-import { useUsers } from '~/features';
+import { useUsers, type UsersSearchParams } from '~/features';
 import {
   Highlight,
   IconButtonLink,
   MenuItemLink,
   OptionalLabel,
-  sliceCollection,
   Table,
   TableCell,
   UserAvatar,
-  type User,
+  type NonUndefined,
 } from '~/shared';
 
-export const usersSortingFields = ['firstName', 'lastName', 'email', 'positionName', 'departmentName'] as const;
-
-function mapSortToProperty(sort: (typeof usersSortingFields)[number]) {
-  return (user: User) => {
-    let result;
-
-    switch (sort) {
-      case 'firstName':
-        result = user.profile.firstName;
-        break;
-      case 'lastName':
-        result = user.profile.lastName;
-        break;
-      case 'email':
-        result = user.email;
-        break;
-      case 'positionName':
-        result = user.positionName;
-        break;
-      case 'departmentName':
-        result = user.departmentName;
-        break;
-    }
-
-    return result?.toLowerCase() ?? '';
-  };
-}
-
-function createCompareFn(sort: (typeof usersSortingFields)[number], order: 'asc' | 'desc') {
-  const getValue = mapSortToProperty(sort);
-  const isAsc = order === 'asc';
-
-  return (a: User, b: User) => {
-    const valA = getValue(a);
-    const valB = getValue(b);
-
-    if (typeof valA === 'string' && typeof valB === 'string') {
-      if (valA === '') return isAsc ? 1 : -1;
-      if (valB === '') return isAsc ? -1 : 1;
-
-      const comparison = valA.localeCompare(valB);
-      return isAsc ? comparison : -comparison;
-    }
-
-    if (valA < valB) return isAsc ? -1 : 1;
-    if (valA > valB) return isAsc ? 1 : -1;
-
-    return 0;
-  };
-}
-
-type UsersTableProps = { q: string; sort: (typeof usersSortingFields)[number] } & Omit<
-  React.ComponentProps<typeof Table>,
-  'data' | 'children' | 'headFields' | 'count' | 'sort'
->;
+type UsersTableProps = {
+  q: NonUndefined<UsersSearchParams['q']>;
+  sort: NonUndefined<UsersSearchParams['sort']>;
+} & Omit<React.ComponentProps<typeof Table>, 'data' | 'children' | 'headFields' | 'count' | 'sort'>;
 
 export const UsersTable: React.FC<UsersTableProps> = (props) => {
   const { sort, page, limit, order, q, ...rest } = props;
   const { t } = useTranslation();
-  const { users } = useUsers();
   const auth = useAuth();
+  const { users, total } = useUsers({ sort, order, page, limit, q });
   const [menuOpen, setMenuOpen] = useState(false);
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-
-  // todo: memo
-  const searchResults = fuzzysort.go(q, users, {
-    all: true,
-    threshold: 0,
-    keys: ['email', 'profile.firstName', 'profile.lastName', 'departmentName', 'positionName'],
-  });
-
-  const searchedUsers = searchResults.map((result) => {
-    const [email, firstName, lastName, department, position] = result;
-
-    return {
-      ...result.obj,
-      highlights: {
-        email: email.highlight(),
-        firstName: firstName.highlight(),
-        lastName: lastName.highlight(),
-        departmentName: department.highlight(),
-        positionName: position.highlight(),
-      },
-    };
-  });
-
-  const sortedUsers = [...searchedUsers].sort(createCompareFn(sort, order));
-
-  const filteredUsers = sortedUsers;
 
   function handleOpenMenu(e: React.MouseEvent<HTMLButtonElement>) {
     setMenuOpen(true);
@@ -120,7 +41,7 @@ export const UsersTable: React.FC<UsersTableProps> = (props) => {
 
   return (
     <Table
-      data={sliceCollection(filteredUsers, { page, limit })}
+      data={users}
       headFields={[
         { id: 'avatar', title: '' },
         { id: 'firstName', title: t('First Name') },
@@ -130,7 +51,7 @@ export const UsersTable: React.FC<UsersTableProps> = (props) => {
         { id: 'positionName', title: t('Position') },
         { id: 'action', title: '' },
       ]}
-      count={filteredUsers.length}
+      count={total}
       page={page}
       limit={limit}
       sort={sort}
